requirements.txt
transformers
datasets
accelerate
faiss-cpu
torch
requests
